+++ src/api/customframes.js
@@ -1,6 +1,6 @@
 import axios from "axios";
 
-const BASE_URL = "http://localhost:8000/api/v1";
+const BASE_URL = `${import.meta.env.VITE_BASE_URL}`;
 
 export const createCustomFrame = async (
   userId,
