+++ src/api/users.js
@@ -1,29 +1,27 @@
+// @ts-nocheck
+
 import axios from "axios";
 
-const BASE_URL = "http://localhost:8000/api/v1";
+const BASE_URL = `${import.meta.env.VITE_BASE_URL}`;
 
 export const login = async (email, password) => {
   try {
-    const response = await axios.post(`${BASE_URL}/users/login/`, {
+    const response = await axios.post(`${BASE_URL}/users/login`, {
       email,
       password,
     });
-
-    const userId = response.data.userId;
-
-    localStorage.setItem("userId", userId);
-
     return response.data;
   } catch (error) {
     return error.response.data;
   }
 };
 
-export const signup = async (email, password) => {
+export const signup = async (email, password, username) => {
   try {
-    const response = await axios.post(`${BASE_URL}/users/sign-up/`, {
+    const response = await axios.post(`${BASE_URL}/users/sign-up`, {
       email,
       password,
+      username,
     });
     return response.data;
   } catch (error) {
