--- src/pages/LoginPage/index.css
@@ -1,68 +1,0 @@
-.page {
-  position: absolute;
-  top: 0;
-  bottom: 0;
-  width: 100%;
-  max-width: 500px;
-  padding: 0 20px;
-
-  left: 50%;
-  transform: translate(-50%, 0);
-
-  background-color: #ffffff;
-
-  overflow: hidden;
-
-  display: flex;
-  flex-direction: column;
-}
-
-.titleWrap {
-  margin-top: 87px;
-  font-size: 20px;
-  font-weight: 700;
-  color: #000;
-}
-
-.contentWrap {
-  margin-top: 26px;
-  flex: 1;
-}
-
-.inputTitle {
-  font-size: 12px;
-  font-weight: 600;
-  color: #000;
-}
-
-.inputWrap {
-  display: flex;
-  border-radius: 10px;
-  padding: 16px;
-  margin-top: 9px;
-  background-color: #ffffff;
-  border: 1px solid #999;
-}
-
-.inputWrap:focus-within {
-  border: 1px solid #8761d2;
-}
-
-.input {
-  width: 100%;
-  outline: none;
-  border: none;
-  height: 17px;
-  font-size: 14px;
-  font-weight: 400;
-}
-
-.input::placeholder {
-  color: #dadada;
-}
-
-.errorMessage {
-  margin-top: 8px;
-  color: #ef0000;
-  font-size: 12px;
-}
