+++ src/routes/routePath.js
@@ -4,6 +4,10 @@
 // import MyCustomFramesPageTest from "@/test/MyCustomFramesPageTest";
 
 const RoutePath = {
+  /* 로그인 페이지 */
+  Login: "/login",
+  Signup: "/signup",
+
   /* 메인 페이지 */
   Main: "/",
   HotFramePage: "/hotframe",
